✨ feat(vectors): add AbstractVectors ABC parent to Spaces
